+++ tests/ops/models/test_privacy_request.py
@@ -17,6 +17,7 @@ from fidesops.ops.models.policy import CurrentStep, Policy
 from fidesops.ops.models.privacy_request import (
     CheckpointActionRequired,
     Consent,
+    ConsentRequest,
     PrivacyRequest,
     PrivacyRequestStatus,
     ProvidedIdentity,
@@ -696,6 +697,33 @@ def test_consent(db):
 
     assert consent_data_1["data_use"] in data_uses
     assert consent_data_2["data_use"] in data_uses
+
+    provided_identity.delete(db)
+
+    assert Consent.get(db, object_id=consent_1.id) is None
+    assert Consent.get(db, object_id=consent_2.id) is None
+
+
+def test_consent_request(db):
+    provided_identity_data = {
+        "privacy_request_id": None,
+        "field_name": "email",
+        "encrypted_value": {"value": "test@email.com"},
+    }
+    provided_identity = ProvidedIdentity.create(db, data=provided_identity_data)
+
+    consent_request_1 = {
+        "provided_identity_id": provided_identity.id,
+    }
+    consent_1 = ConsentRequest.create(db, data=consent_request_1)
+
+    consent_request_2 = {
+        "provided_identity_id": provided_identity.id,
+    }
+    consent_2 = ConsentRequest.create(db, data=consent_request_2)
+
+    assert consent_1.provided_identity_id in provided_identity.id
+    assert consent_2.provided_identity_id in provided_identity.id
 
     provided_identity.delete(db)